move TrackData to constant.dart
