+++ lib/src/Evaluator/feature_evaluator.dart
@@ -3,7 +3,6 @@ import 'dart:developer';
 import 'package:growthbook_sdk_flutter/growthbook_sdk_flutter.dart';
 import 'package:growthbook_sdk_flutter/src/Evaluator/experiment_helper.dart';
 import 'package:growthbook_sdk_flutter/src/MultiUserMode/Model/evaluation_context.dart';
-import 'package:growthbook_sdk_flutter/src/Utils/gb_variation_meta.dart';
 
 /// Feature Evaluator Class
 /// Takes Context and Feature Key
