+++ lib/src/MultiUserMode/constant.dart
@@ -1,5 +1,4 @@
 import 'package:growthbook_sdk_flutter/growthbook_sdk_flutter.dart';
-import 'package:growthbook_sdk_flutter/src/Utils/gb_variation_meta.dart';
 
 typedef TrackingCallBackWithUser = void Function(GBTrackData)?;
