+++ lib/src/growth_book_sdk.dart
@@ -8,7 +8,6 @@ import 'package:growthbook_sdk_flutter/src/Model/sticky_assignments_document.dar
 import 'package:growthbook_sdk_flutter/src/MultiUserMode/Model/evaluation_context.dart';
 import 'package:growthbook_sdk_flutter/src/StickyBucketService/sticky_bucket_service.dart';
 import 'package:growthbook_sdk_flutter/src/Utils/crypto.dart';
-import 'package:growthbook_sdk_flutter/src/Utils/gb_variation_meta.dart';
 
 typedef VoidCallback = void Function();
