+++ pubspec.yaml
@@ -13,15 +13,15 @@ dependencies:
   flutter:
     sdk: flutter
   json_annotation: ^4.5.0
-  pointycastle: ^3.1.3
+  pointycastle: ^4.0.0
   shared_preferences: ^2.2.3
   tuple: ^2.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^3.0.1
-  build_runner: ^2.1.11
-  analyzer: ^6.2.0
-  json_serializable: ^6.7.1
+  flutter_lints: ^5.0.0
+  build_runner: any
+  analyzer: ^7.3.0
+  json_serializable: any
